gfxlib2: Remove int cast, just in case X11 window id can be 64bit

+++ src/gfxlib2/unix/gfx_x11.c
@@ -849,7 +849,7 @@ int fb_hX11ScreenInfo(int *width, int *height, int *depth, int *refresh)
 
 ssize_t fb_hGetWindowHandle(void)
 {
-	return (fb_x11.display ? (int)fb_x11.window : 0);
+	return (fb_x11.display ? fb_x11.window : 0);
 }
 
 #else
